+++ common/src/main/java/com/st0x0ef/stellaris/common/menus/RadioactiveGeneratorMenu.java
@@ -112,7 +112,7 @@ public class RadioactiveGeneratorMenu extends AbstractContainerMenu {
     public void syncBattery(ServerPlayer player) {
         if (!player.level().isClientSide()) {
 
-            NetworkManager.sendToPlayer(player, new SyncWidgetsTanksPacket(new int[] {this.getBlockEntity().getEnergy(null).getEnergy()}
+            NetworkManager.sendToPlayer(player, new SyncWidgetsTanksPacket(new long[] {this.getBlockEntity().getEnergy(null).getEnergy()}
             ));
         }
     }
